Strip port from url of https requests [gh-1]
Added specs to ensure the behavior at:

  https://www.twilio.com/docs/security

under the 'A few notes' section is held to.  The new spec to check
that credentials are not considered in the middleware when validating
passed on the first try; upon checking Rack::Request is not including
them in the url.  The environment doesn't put them in the HTTP_HOST
which Rack is pulling from, rather, they are put into basic auth in
the HTTP_AUTHORIZATION field